Restructured the files and corrected errors
