+++ src/main/java/seedu/drugs/DrugAction.java
@@ -1,4 +1,4 @@
-package seedu.duke;
+package seedu.drugs;
 
 import java.util.ArrayList;
