+++ src/main/java/seedu/duke/exceptions/DukeException.java
@@ -1,4 +1,4 @@
-package seedu.duke;
+package seedu.duke.exceptions;
 
 public class DukeException extends Exception {
