+++ src/main/java/seedu/staff/Doctor.java
@@ -1,4 +1,4 @@
-package system.staff;
+package seedu.staff;
 
 public class Doctor extends Staff{
