+++ src/main/java/seedu/staff/UI.java
@@ -1,4 +1,4 @@
-package system.staff;
+package seedu.staff;
 
 public class UI {
